src/app/second-route-component/second-route-component.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { SecondRouteComponent } from './second-route-component.component';

describe('Component: SecondRouteComponent', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [SecondRouteComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([SecondRouteComponent],
      (component: SecondRouteComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(SecondRouteComponentComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(SecondRouteComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-second-route-component></app-second-route-component>
  `,
  directives: [SecondRouteComponent]
})
class SecondRouteComponentComponentTestController {
}

